week 2 wireframes
For my project, I will be designing a Web site that is a
e-commerce/blog hybrid called “Curated Goods.” The goal of the site is
to showcase an everchanging display of up and coming trends and new
ideas. Categories will include food, music, design, home and style. In
my first wireframe, the focal point is a diamond grid of clickable
images which will represent each category of curated goods. It will
bring readers to another page of a grid of curated goods for that
category, which they can read about each good, who designed it,
inspiration, as well an opportunity to purchase said item. There will
also be a separate blog that will serve as the more “newsy” section
where it will place more of a spotlight on the makers, new openings,
lines, etc.,  My second wireframe treats the page as a scroll with a
hero image as the focal point. This hero image will direct readers to
the blog, but they can also continue to scroll down to the diamond grid
of categories which viewers can see as it peeps beyond the top of the
fold. The logo in this version is smaller to avoid competing with the
hero image and the links to navigate the site are diamond shaped icon
links to create continuity in the design by mimicking the diamond grid
below the fold.